+++ site/src/components/layout/responsive-footer.tsx
